Refactor code for cleaner organization
