+++ languageserver/src/main/scala/scala/meta/lsp/CancelParams.scala
@@ -1,0 +1,11 @@
+package scala.meta.lsp
+
+import play.api.libs.json.JsValue
+import play.api.libs.json.Json
+import play.api.libs.json.OFormat
+
+case class CancelParams(id: JsValue)
+
+object CancelParams {
+  implicit val format: OFormat[CancelParams] = Json.format[CancelParams]
+}
