+++ languageserver/src/main/scala/scala/meta/lsp/ErrorObject.scala
@@ -1,0 +1,11 @@
+package scala.meta.lsp
+
+import play.api.libs.json.JsValue
+import play.api.libs.json.Json
+import play.api.libs.json.OFormat
+
+case class ErrorObject(code: ErrorCode, message: String, data: Option[JsValue])
+
+object ErrorObject {
+  implicit val format: OFormat[ErrorObject] = Json.format[ErrorObject]
+}
